refactor: created a folder for mongo related components
