+++ .gitignore
@@ -131,3 +131,4 @@ dmypy.json
 __SUCCESS__
 
 flycheck*
+.vscode/
