+++ scribe/mongo_scribe/mongo/command/commands.py
