+++ scribe/mongo_scribe/mongo/command/commons.py
