+++ scribe/mongo_scribe/mongo/command/decode.py
