+++ scribe/mongo_scribe/mongo/command/exceptions.py
