+++ scribe/mongo_scribe/mongo/db/connection.py
