+++ scribe/mongo_scribe/mongo/db/executor.py
